+++ organization/src/main/java/org/exoplatform/management/organization/user/UserImportResource.java
@@ -152,7 +152,7 @@ public class UserImportResource implements OperationHandler {
     } catch (Exception e) {
       throw new OperationException(OperationNames.IMPORT_RESOURCE, "Error while reading View Templates from Stream.", e);
     } finally {
-      if (tempFile != null) {
+      if (tempFile != null && tempFile.exists()) {
         try {
           tempFile.delete();
         } catch (Exception e) {
